requirements.txt
discord.py==0.16.8
oauth2client==4.1.2
psycopg2-binary==2.8